Add stories viewers and likes, Posts likes and comments, Comments replies and likes, and improve a lot of things (including the clean architecture)

+++ Server - Clean Architecture/Use_cases/comment/delete_comment_by_id.js
@@ -5,6 +5,10 @@ function buildDeleteCommentById({ CommentsDB, Id, AppError }) {
             throw new AppError("Can't delete comment by invalid id.")
         }
 
+        if (!(await CommentsDB.doesCommentExist(commentId))) {
+            throw new AppError("Comment doesn't exist.")
+        }
+
         await CommentsDB.deleteById(commentId)
     }
 }
