+++ Server - Clean Architecture/Use_cases/post/delete_post_by_id.js
@@ -5,6 +5,10 @@ function buildDeletePostById({ PostsDB, Id, AppError }) {
             throw new AppError("Can't delete post by invalid id.")
         }
 
+        if (!(await PostsDB.doesPostExist(postId))) {
+            throw new AppError("Post doesn't exist.")
+        }
+
         await PostsDB.deleteById(postId)
     }
 }
